+++ src/perceptron.rs
@@ -3,10 +3,10 @@ use crate::iris::Iris;
 #[allow(dead_code)]
 #[derive(Debug, Clone)]
 pub struct Perceptron {
-    learning_rate: f64,
+    learning_rate: f32,
     epochs: u32,
-    weights: Vec<Vec<f64>>,
-    bias: f64,
+    weights: Vec<Vec<f32>>,
+    bias: f32,
     error_count: Vec<u32>,
 }
 
@@ -14,7 +14,7 @@ pub struct Perceptron {
 impl Perceptron {
     pub fn new(learning_rate: Option<f64>, epochs: Option<u32>, input: &Vec<Iris>) -> Self {
         Perceptron {
-            learning_rate: learning_rate.unwrap_or(0.1),
+            learning_rate: learning_rate.unwrap_or(0.1) as f32,
             epochs: epochs.unwrap_or(10),
             weights: vec![vec![1e-4; 4]; input.len()],
             bias: 0.0,
@@ -22,7 +22,7 @@ impl Perceptron {
         }
     }
 
-    pub fn fit(&mut self, X: &Vec<Iris>, y: &Vec<f64>) -> Self {
+    pub fn fit(&mut self, X: &Vec<Iris>, y: &Vec<f32>) -> Self {
         let mut update;
         let mut errors = 0;
         let mut weight_index = 0;
@@ -49,7 +49,7 @@ impl Perceptron {
         self.clone()
     }
 
-    pub fn predict(&self, wi: usize, X: &Iris) -> f64 {
+    pub fn predict(&self, wi: usize, X: &Iris) -> f32 {
         let output = self.net_input(wi, X);
         match output >= 0.0 {
             true => 1.0,
@@ -57,13 +57,13 @@ impl Perceptron {
         }
     }
 
-    fn net_input(&self, wi: usize, X: &Iris) -> f64 {
+    fn net_input(&self, wi: usize, X: &Iris) -> f32 {
         let output;
         output = self.dot(wi, X) + self.bias;
         output
     }
 
-    fn update_weights(&mut self, wi: usize, X: &Iris, update: f64) {
+    fn update_weights(&mut self, wi: usize, X: &Iris, update: f32) {
         let weights = &mut self.weights;
         for i in 0..X.len() {
             weights[wi][i] += update * X[i];
@@ -74,7 +74,7 @@ impl Perceptron {
         self.error_count.clone()
     }
 
-    fn dot(&self, wi: usize, X: &Iris) -> f64 {
+    fn dot(&self, wi: usize, X: &Iris) -> f32 {
         let mut output = 0.0;
         for i in 0..X.len() {
             output += self.weights[wi][i] * X[i];
